add values and for loop
Added the suits and values array that will contain the values and the
suits required for the cards. Used a for loop to loop through these and
get the list of all possible combinations that are available. Used the
format macros to join and format the list of combinations that is being
generated.